--- a/flow-typed/recast.js
+++ b/flow-typed/recast.js
@@ -9,7 +9,7 @@
 /*eslint no-unused-vars: 0*/
 
 /**
- * A minimal set of declarations to make flow work with the recast API.
+ * A minimal set of declarations to make flow work with the ast-types API.
  */
 
 type ASTNode = Object;
@@ -34,9 +34,7 @@ declare class NodePath {
   map<T>(f: (p: NodePath) => T): Array<T>;
   filter(f: (p: NodePath) => boolean): Array<NodePath>;
   push(node: ASTNode): void;
-}
 
-type Recast = {
-  parse: (src: string) => ASTNode,
-  print: (path: NodePath) => { code: string },
-};
+  // our addition
+  getSrc(): string;
+}