--- a/src/handlers/componentDocblockHandler.js
+++ b/src/handlers/componentDocblockHandler.js
@@ -7,20 +7,15 @@
  * @flow
  */
 
+import types from 'ast-types';
 import type Documentation from '../Documentation';
-
-import recast from 'recast';
 import { getDocblock } from '../utils/docblock';
 
-const {
-  types: { namedTypes: types },
-} = recast;
+const { namedTypes: t } = types;
 
 function isClassDefinition(nodePath) {
   const node = nodePath.node;
-  return (
-    types.ClassDeclaration.check(node) || types.ClassExpression.check(node)
-  );
+  return t.ClassDeclaration.check(node) || t.ClassExpression.check(node);
 }
 
 /**
@@ -45,14 +40,14 @@ export default function componentDocblockHandler(
   if (description == null) {
     // Find parent statement (e.g. var Component = React.createClass(<path>);)
     let searchPath = path;
-    while (searchPath && !types.Statement.check(searchPath.node)) {
+    while (searchPath && !t.Statement.check(searchPath.node)) {
       searchPath = searchPath.parent;
     }
     if (searchPath) {
       // If the parent is an export statement, we have to traverse one more up
       if (
-        types.ExportNamedDeclaration.check(searchPath.parentPath.node) ||
-        types.ExportDefaultDeclaration.check(searchPath.parentPath.node)
+        t.ExportNamedDeclaration.check(searchPath.parentPath.node) ||
+        t.ExportDefaultDeclaration.check(searchPath.parentPath.node)
       ) {
         searchPath = searchPath.parentPath;
       }