--- a/src/handlers/displayNameHandler.js
+++ b/src/handlers/displayNameHandler.js
@@ -7,17 +7,14 @@
  * @flow
  */
 
-import type Documentation from '../Documentation';
-
+import types from 'ast-types';
 import getMemberValuePath from '../utils/getMemberValuePath';
 import getNameOrValue from '../utils/getNameOrValue';
-import recast from 'recast';
 import resolveToValue from '../utils/resolveToValue';
 import resolveFunctionDefinitionToReturnValue from '../utils/resolveFunctionDefinitionToReturnValue';
+import type Documentation from '../Documentation';
 
-const {
-  types: { namedTypes: types },
-} = recast;
+const { namedTypes: t } = types;
 
 export default function displayNameHandler(
   documentation: Documentation,
@@ -28,27 +25,27 @@ export default function displayNameHandler(
     // Function and class declarations need special treatment. The name of the
     // function / class is the displayName
     if (
-      types.ClassDeclaration.check(path.node) ||
-      types.FunctionDeclaration.check(path.node)
+      t.ClassDeclaration.check(path.node) ||
+      t.FunctionDeclaration.check(path.node)
     ) {
       documentation.set('displayName', getNameOrValue(path.get('id')));
     } else if (
-      types.ArrowFunctionExpression.check(path.node) ||
-      types.FunctionExpression.check(path.node)
+      t.ArrowFunctionExpression.check(path.node) ||
+      t.FunctionExpression.check(path.node)
     ) {
       let currentPath = path;
       while (currentPath.parent) {
-        if (types.VariableDeclarator.check(currentPath.parent.node)) {
+        if (t.VariableDeclarator.check(currentPath.parent.node)) {
           documentation.set(
             'displayName',
             getNameOrValue(currentPath.parent.get('id')),
           );
           return;
-        } else if (types.AssignmentExpression.check(currentPath.parent.node)) {
+        } else if (t.AssignmentExpression.check(currentPath.parent.node)) {
           const leftPath = currentPath.parent.get('left');
           if (
-            types.Identifier.check(leftPath.node) ||
-            types.Literal.check(leftPath.node)
+            t.Identifier.check(leftPath.node) ||
+            t.Literal.check(leftPath.node)
           ) {
             documentation.set('displayName', getNameOrValue(leftPath));
             return;
@@ -64,10 +61,10 @@ export default function displayNameHandler(
   // If display name is defined as a getter we get a function expression as
   // value. In that case we try to determine the value from the return
   // statement.
-  if (types.FunctionExpression.check(displayNamePath.node)) {
+  if (t.FunctionExpression.check(displayNamePath.node)) {
     displayNamePath = resolveFunctionDefinitionToReturnValue(displayNamePath);
   }
-  if (!displayNamePath || !types.Literal.check(displayNamePath.node)) {
+  if (!displayNamePath || !t.Literal.check(displayNamePath.node)) {
     return;
   }
   documentation.set('displayName', displayNamePath.node.value);