--- a/src/handlers/flowTypeHandler.js
+++ b/src/handlers/flowTypeHandler.js
@@ -6,10 +6,8 @@
  *
  * @flow
  */
-import recast from 'recast';
 
-import type Documentation from '../Documentation';
-
+import types from 'ast-types';
 import getFlowType from '../utils/getFlowType';
 import getTSType from '../utils/getTSType';
 import getPropertyName from '../utils/getPropertyName';
@@ -20,19 +18,19 @@ import resolveToValue from '../utils/resolveToValue';
 import setPropDescription from '../utils/setPropDescription';
 import { unwrapUtilityType } from '../utils/flowUtilityTypes';
 import { type TypeParameters } from '../utils/getTypeParameters';
+import type Documentation from '../Documentation';
 
-const {
-  types: { namedTypes: types },
-} = recast;
+const { namedTypes: t } = types;
+
 function setPropDescriptor(
   documentation: Documentation,
   path: NodePath,
   typeParams: ?TypeParameters,
 ): void {
-  if (types.ObjectTypeSpreadProperty.check(path.node)) {
+  if (t.ObjectTypeSpreadProperty.check(path.node)) {
     const argument = unwrapUtilityType(path.get('argument'));
 
-    if (types.ObjectTypeAnnotation.check(argument.node)) {
+    if (t.ObjectTypeAnnotation.check(argument.node)) {
       applyToFlowTypeProperties(
         documentation,
         argument,
@@ -47,7 +45,7 @@ function setPropDescriptor(
     const name = argument.get('id').get('name');
     const resolvedPath = resolveToValue(name);
 
-    if (resolvedPath && types.TypeAlias.check(resolvedPath.node)) {
+    if (resolvedPath && t.TypeAlias.check(resolvedPath.node)) {
       const right = resolvedPath.get('right');
       applyToFlowTypeProperties(
         documentation,
@@ -60,7 +58,7 @@ function setPropDescriptor(
     } else {
       documentation.addComposes(name.node.name);
     }
-  } else if (types.ObjectTypeProperty.check(path.node)) {
+  } else if (t.ObjectTypeProperty.check(path.node)) {
     const type = getFlowType(path.get('value'), typeParams);
     const propName = getPropertyName(path);
     if (!propName) return;
@@ -73,7 +71,7 @@ function setPropDescriptor(
     // to not need to duplicate the logic for checking for
     // imported types that are spread in to props.
     setPropDescription(documentation, path);
-  } else if (types.TSPropertySignature.check(path.node)) {
+  } else if (t.TSPropertySignature.check(path.node)) {
     const type = getTSType(path.get('typeAnnotation'), typeParams);
 
     const propName = getPropertyName(path);