--- a/src/resolver/findAllExportedComponentDefinitions.js
+++ b/src/resolver/findAllExportedComponentDefinitions.js
@@ -7,6 +7,7 @@
  * @flow
  */
 
+import types from 'ast-types';
 import isExportsOrModuleAssignment from '../utils/isExportsOrModuleAssignment';
 import isReactComponentClass from '../utils/isReactComponentClass';
 import isReactCreateClassCall from '../utils/isReactCreateClassCall';
@@ -17,11 +18,13 @@ import resolveExportDeclaration from '../utils/resolveExportDeclaration';
 import resolveToValue from '../utils/resolveToValue';
 import resolveHOC from '../utils/resolveHOC';
 
-function ignore() {
+const { visit, namedTypes: t } = types;
+
+function ignore(): false {
   return false;
 }
 
-function isComponentDefinition(path) {
+function isComponentDefinition(path: NodePath): boolean {
   return (
     isReactCreateClassCall(path) ||
     isReactComponentClass(path) ||
@@ -30,11 +33,11 @@ function isComponentDefinition(path) {
   );
 }
 
-function resolveDefinition(definition, types): ?NodePath {
+function resolveDefinition(definition: NodePath): ?NodePath {
   if (isReactCreateClassCall(definition)) {
     // return argument
     const resolvedPath = resolveToValue(definition.get('arguments', 0));
-    if (types.ObjectExpression.check(resolvedPath.node)) {
+    if (t.ObjectExpression.check(resolvedPath.node)) {
       return resolvedPath;
     }
   } else if (isReactComponentClass(definition)) {
@@ -66,13 +69,11 @@ function resolveDefinition(definition, types): ?NodePath {
  */
 export default function findExportedComponentDefinitions(
   ast: ASTNode,
-  recast: Object,
 ): Array<NodePath> {
-  const types = recast.types.namedTypes;
   const components: Array<NodePath> = [];
 
-  function exportDeclaration(path) {
-    const definitions: Array<?NodePath> = resolveExportDeclaration(path, types)
+  function exportDeclaration(path: NodePath): ?boolean {
+    const definitions: Array<?NodePath> = resolveExportDeclaration(path)
       .reduce((acc, definition) => {
         if (isComponentDefinition(definition)) {
           acc.push(definition);
@@ -84,7 +85,7 @@ export default function findExportedComponentDefinitions(
         }
         return acc;
       }, [])
-      .map(definition => resolveDefinition(definition, types));
+      .map(definition => resolveDefinition(definition));
 
     if (definitions.length === 0) {
       return false;
@@ -97,7 +98,7 @@ export default function findExportedComponentDefinitions(
     return false;
   }
 
-  recast.visit(ast, {
+  visit(ast, {
     visitFunctionDeclaration: ignore,
     visitFunctionExpression: ignore,
     visitClassDeclaration: ignore,
@@ -115,7 +116,7 @@ export default function findExportedComponentDefinitions(
     visitExportNamedDeclaration: exportDeclaration,
     visitExportDefaultDeclaration: exportDeclaration,
 
-    visitAssignmentExpression: function(path) {
+    visitAssignmentExpression: function(path: NodePath): ?boolean {
       // Ignore anything that is not `exports.X = ...;` or
       // `module.exports = ...;`
       if (!isExportsOrModuleAssignment(path)) {
@@ -130,7 +131,7 @@ export default function findExportedComponentDefinitions(
           return false;
         }
       }
-      const definition = resolveDefinition(path, types);
+      const definition = resolveDefinition(path);
       if (definition && components.indexOf(definition) === -1) {
         components.push(definition);
       }