--- a/src/utils/resolveObjectValuesToArray.js
+++ b/src/utils/resolveObjectValuesToArray.js
@@ -122,7 +122,7 @@ export function resolveObjectToPropMap(
 }
 
 /**
- * Returns an ArrayExpression which contains all the keys resolved from an object
+ * Returns an ArrayExpression which contains all the values resolved from an object
  *
  * Ignores setters in objects
  *