… README.md
… benchmark/fixtures/CircularProgress.js
… benchmark/index.js
… bin/__tests__/example/customResolver.js
… bin/__tests__/react-docgen-test.js
… package.json
… src/__tests__/main-test.js
… src/__tests__/parse-test.js
… src/handlers/__tests__/componentDocblockHandler-test.js
… src/handlers/__tests__/componentMethodsHandler-test.js
… src/handlers/__tests__/defaultPropsHandler-test.js
… src/handlers/__tests__/displayNameHandler-test.js
… src/handlers/__tests__/flowTypeHandler-test.js
… src/handlers/__tests__/propDocblockHandler-test.js
… src/handlers/__tests__/propTypeCompositionHandler-test.js
… src/resolver/__tests__/findAllComponentDefinitions-test.js
… src/resolver/__tests__/findAllExportedComponentDefinitions-test.js
… src/resolver/__tests__/findExportedComponentDefinition-test.js
… src/resolver/findExportedComponentDefinition.js
… src/types.js
… src/utils/__tests__/__snapshots__/getMembers-test.js.snap
… src/utils/__tests__/__snapshots__/getPropType-test.js.snap
… src/utils/__tests__/docblock-test.js
… src/utils/__tests__/flowUtilityTypes-test.js
… src/utils/__tests__/getClassMemberValuePath-test.js
… src/utils/__tests__/getFlowType-test.js
… src/utils/__tests__/getMemberExpressionRoot-test.js
… src/utils/__tests__/getMemberExpressionValuePath-test.js
… src/utils/__tests__/getMemberValuePath-test.js
… src/utils/__tests__/getMembers-test.js