This removes recast and only uses ast-types directly.

The performance increases significantly, but I haven't tested more different cases. 

There are still some cases that do not yet work:

- [x] Printing of nodes which were created in react-docgen by ast-types builders.
- [x] Comments in printed code are not filtered out

Fixes #237