@fkling 2019-04-23
Amazing. Somehow I thought the whole `NodePath` business was part of recast (that would probably be more difficult to remove).

Looking forward to where this goes 😃 

> The performance increases significantly, but I haven't tested more different cases.

Can you share numbers?