@danez 2019-04-23
yeah I forgot to include that yesterday:

Before:
```
❯ node --expose-gc --max-old-space-size=2000 --max-semi-space-size=1500 --noconcurrent_sweeping ./benchmark/index.js
Node: v10.15.1
Running benchmark for CircularProgress.js ...
┌─────────────────────┬─────────────────────────────┐
│ fixture             │ timing                      │
├─────────────────────┼─────────────────────────────┤
│ CircularProgress.js │ 30.34 ops/sec ±7.82% (33ms) │
└─────────────────────┴─────────────────────────────┘
```

With PR:
```
❯ node --expose-gc --max-old-space-size=2000 --max-semi-space-size=1500 --noconcurrent_sweeping ./benchmark/index.js
Node: v10.15.1
Running benchmark for CircularProgress.js ...
┌─────────────────────┬─────────────────────────────┐
│ fixture             │ timing                      │
├─────────────────────┼─────────────────────────────┤
│ CircularProgress.js │ 40.55 ops/sec ±5.08% (25ms) │
└─────────────────────┴─────────────────────────────┘
```

> The commandline options are to prevent gc to interfere, so the script runs without automatic gc.

The test file is taken from material ui. I need to find bigger or different (flow) files to test on. Would also be interesting to run this on a big codebase and see the difference.