@fkling 2019-04-23
>  Would also be interesting to run this on a big codebase and see the difference.

I get the hint 😉